Fix #52: enable to use constant CAN_BUFFER_NUMBER in preprocessor

+++ MDR32F9Qx_StdPeriph_Driver/inc/MDR32F9Qx_can.h
@@ -376,7 +376,7 @@ typedef struct
   * @{
   */
 
-#define CAN_BUFFER_NUMBER           ((uint32_t)32)          /*!< Total buffer number */
+#define CAN_BUFFER_NUMBER           (32u)          /*!< Total buffer number */
 
 /** @} */ /* End of group CAN_Constants */
